requirements.txt
streamlit
pandas
beepy
plotly
google-generativeai # Added for Gemini API